+++ constant/prefix.go
@@ -1,6 +1,0 @@
-package constant
-
-const (
-	TempPrefix  = Mangal + "Temp_"
-	CachePrefix = Mangal + "Cache_"
-)