+++ provider/mangadex/cache.go
@@ -2,7 +2,9 @@ package mangadex
 
 import (
 	"github.com/metafates/mangal/cache"
+	"github.com/metafates/mangal/where"
 	"github.com/samber/mo"
+	"path/filepath"
 	"time"
 )
 
@@ -13,7 +15,7 @@ type cacher[T any] struct {
 func newCacher[T any](name string) *cacher[T] {
 	return &cacher[T]{
 		internal: cache.New[map[string]T](
-			name,
+			filepath.Join(where.Cache(), name+".json"),
 			&cache.Options{
 				ExpireEvery: mo.Some(time.Hour * 24),
 			},